backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
spacy
sentence-transformers
librosa
soundfile
numpy
python-multipart
opencv-python-headless
mediapipe
ultralytics
websockets
sqlalchemy
passlib[bcrypt]
pyjwt
pymongo
